+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.3.31
 
+* [X] #121 Complex enums only accept `enum_name = scalar`, not `enum_name = bool` or `enum_name = int`
 * [x] #122 Non-Stellaris games raise warning if modifier icon doesn't exist
 
 ## 1.3.30